+++ CHANGELOG.md
@@ -1,3 +1,29 @@
-## 0.0.1
+# Changelog
 
-* TODO: Describe initial release.
+## 0.2.0
+
+* Added timestamp to events
+
+## 0.1.6
+
+* Fixed delay problem that avoided events to be reported (#3).
+
+## 0.1.5
+
+* Fixed compilation error (#2). Added support for new embedding 
+mechanism.
+
+## 0.1.2
+
+* Updated Android dependencies
+
+## 0.1.1
+
+* Updated description.
+* Formatted code.
+
+## 0.1.0
+
+* Sensor stream to retrieve data.
+* Dynamic sensor registering in Android.
+* iOS register some specific sensors. (Check Sensors.dart class)